refactor: Move codemaker mock into a place that tsc is happy with
