+++ src/utils/snapshot.test.ts
@@ -1,4 +1,4 @@
-import { MockCodeMaker } from "../../test/utils/codemaker";
+import { MockCodeMaker } from "./codemaker";
 import { Imports } from "./imports";
 import { TestBuilder } from "./snapshot";
